Refactor gRPC client fixture to conftest.py and remove duplicates from integration tests
