+++ tests/integration/test_save_to_data_platform.py
@@ -4,9 +4,7 @@ from importlib.metadata import version
 import numpy as np
 import pandas as pd
 import pytest
-import pytest_asyncio
 from betterproto.lib.google.protobuf import Struct, Value
-from grpclib.client import Channel
 from ocf import dp
 
 from src.pvnet_app.save import (
@@ -14,20 +12,6 @@ from src.pvnet_app.save import (
     limit_adjuster,
     save_forecast_to_data_platform,
 )
-
-
-# @pytest.fixture(scope="session")
-@pytest_asyncio.fixture(scope="session")
-async def client(dp_client):
-    """
-    Fixture to create a gRPC client connected to the shared Data Platform server.
-    """
-    host, port = dp_client
-    channel = Channel(host=host, port=port)
-    client_stub = dp.DataPlatformDataServiceStub(channel)
-
-    yield client_stub
-    channel.close()
 
 
 @pytest.mark.asyncio(loop_scope="session")
